backend/pyproject.toml
[project]
name = "openpa"
version = "0.1.0"
description = "OpenPA — Personal Assistant-as-a-Service with LLM agents and MCP tools"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.89.0",
    "bcrypt>=5.0.0",
    "ddgs>=9.13.0",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "feedparser>=6.0.12",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "google-genai>=1.70.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openai>=2.30.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "telethon>=1.42.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
